seperate k8sexecutor classes and add type hinting
